reinvent_2024_agentic/lambda_functions/gen_aws_diag_docker/requirements.txt
boto3
pillow
diagrams